+++ src/App.css
@@ -1,8 +1,9 @@
 *,
-:after,
-:before {
-  margin: 0;
+*::before,
+*::after {
   box-sizing: border-box;
+  padding: 0;
+  margin: 0;
 }
 
 p,
@@ -13,4 +14,9 @@ h4,
 h5,
 h6 {
   margin: 5px;
+}
+
+a {
+  text-decoration: none;
+  color: inherit;
 }
