--- src/components/Footer/Footer.css
@@ -1,114 +1,0 @@
-#menu-bar {
-  width: 30px;
-  height: 30px;
-  cursor: pointer;
-}
-
-.bar {
-  position: relative;
-  height: 4px;
-  width: 100%;
-  background-color: #dc052d;
-  display: block;
-  border-radius: 5px;
-  transition: 0.3s ease;
-  top: 9px;
-}
-
-#bar1 {
-  transform: translateY(-4px);
-}
-
-#bar3 {
-  transform: translateY(4px);
-}
-
-.open {
-  display: block;
-}
-
-.open .bar {
-  background-color: white;
-}
-
-.open #bar1 {
-  transform: translateY(4px) rotateZ(-45deg);
-}
-
-.open #bar2 {
-  opacity: 0;
-}
-
-.open #bar3 {
-  transform: translateY(-4px) rotateZ(45deg);
-}
-.open #bar4 {
-  transform: translateY(4px) rotateZ(225deg);
-}
-
-.open #bar5 {
-  transform: translateY(0px) rotateZ(135deg);
-}
-
-#add-bar {
-  width: 30px;
-  height: 30px;
-  cursor: pointer;
-}
-
-#bar4 {
-  transform: translateY(4px) rotateZ(90deg);
-}
-
-#bar5 {
-  transform: translateY(0px);
-}
-
-.footer-container {
-  border: 1px solid red;
-  display: flex;
-  justify-content: space-between;
-  align-items: stretch;
-  align-content: stretch;
-}
-
-.footer-left {
-  display: flex;
-  align-items: stretch;
-}
-
-.search {
-  margin: auto;
-  max-width: 300px;
-}
-
-form.search input[type='text'] {
-  padding: 6px;
-  font-size: 14px;
-  border: 1px solid grey;
-  float: left;
-  width: 80%;
-  background: #f1f1f1;
-}
-
-form.search button {
-  float: left;
-  width: 14%;
-  padding: 6px;
-  background: red;
-  color: white;
-  font-size: 14px;
-  border: 1px solid grey;
-  border-left: none;
-  cursor: pointer;
-}
-
-form.search button:hover {
-  background: white;
-}
-
-form.search::after {
-  content: '';
-  clear: both;
-  display: table;
-}
